+++ compile.sh
@@ -3,3 +3,4 @@ mkdir build
 cd build
 cmake ..
 make
+./test/unittest
